dashboard/requirements.txt
streamlit
plotly
SQLAlchemy
pandas
folium
streamlit-folium
psycopg2-binary